--- harmony-fe/src/contexts/mobile.ts
@@ -1,3 +1,0 @@
-import React from 'react'
-
-export const MobileContext = React.createContext<boolean>(false)
