--- harmony-fe/src/contexts/user.ts
@@ -1,4 +1,0 @@
-import React from 'react'
-import { User } from '../models/user'
-
-export const UserContext = React.createContext<User | undefined>(undefined)
