+++ harmony-fe/src/pages/home/home.tsx
@@ -2,7 +2,7 @@ import { Container } from '@mui/material'
 import { useContext, useState } from 'react'
 import FilterPostButtons from '../../components/filter-post-buttons'
 import PostFeed from '../../components/post-feed/postFeed'
-import { UserContext } from '../../contexts/user'
+import { UserContext } from '../../contexts/userContext'
 import Head from 'next/head'
 
 export enum FeedMode {
